Revert "Updating tree heights (#975)" (#1032)
This reverts commit 68b0fd844533f5cad46698f12193e0c97e39977d.

+++ circuits/cpp/src/aztec3/circuits/kernel/private/testing_harness.hpp
@@ -15,7 +15,6 @@
 #include <barretenberg/barretenberg.hpp>
 
 #include <array>
-#include <cstdint>
 
 namespace {
 
@@ -48,9 +47,9 @@ using aztec3::circuits::compute_empty_sibling_path;
 constexpr size_t MAX_FUNCTION_LEAVES = 1 << aztec3::FUNCTION_TREE_HEIGHT;  // 2^(height-1)
 // NOTE: *DO NOT* call hashes in static initializers and assign them to constants. This will fail. Instead, use
 // lazy initialization or functions. Lambdas were introduced here.
-const auto EMPTY_FUNCTION_LEAF = [] { return FunctionLeafPreimage<NT>{}.hash(); };           // hash of empty/0 preimage
-const auto EMPTY_CONTRACT_LEAF = [] { return NewContractData<NT>{}.hash(); };                // hash of empty/0 preimage
-constexpr uint64_t PRIVATE_DATA_TREE_NUM_LEAVES = 1ULL << aztec3::PRIVATE_DATA_TREE_HEIGHT;  // 2^(height-1)
+const auto EMPTY_FUNCTION_LEAF = [] { return FunctionLeafPreimage<NT>{}.hash(); };      // hash of empty/0 preimage
+const auto EMPTY_CONTRACT_LEAF = [] { return NewContractData<NT>{}.hash(); };           // hash of empty/0 preimage
+constexpr size_t PRIVATE_DATA_TREE_NUM_LEAVES = 1 << aztec3::PRIVATE_DATA_TREE_HEIGHT;  // 2^(height-1)
 
 inline const auto& get_empty_function_siblings()
 {
